requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.3.0
bcrypt==5.0.0
blis==1.3.0
boto3==1.36.1
botocore==1.36.1
catalogue==2.0.10
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==46.0.3
cymem==2.0.11
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.128.0
fastapi-cli==0.0.7
fastapi-utils==0.8.0
frozenlist==1.7.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
mangum==0.19.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mongoengine==0.29.1
multidict==6.5.0
murmurhash==1.0.13
mypy_extensions==1.1.0
numpy==2.3.1
openai==1.91.0
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pillow==12.1.0
preshed==3.0.10
propcache==0.3.2
psutil==5.9.8
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.2.0
pydantic_core==2.41.5
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
qrcode==8.0
reportlab==4.4.7
requests==2.32.4
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9.1
s3transfer==0.11.1
scikit-learn==1.7.0
scipy==1.16.0
shellingham==1.5.4
six==1.17.0
slugify==0.0.1
smart_open==7.3.0.post1
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.50.0
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.10.0
unicode==2.9
unicode-slugify==0.1.5
Unidecode==1.4.0
urllib3==2.3.0
uvicorn==0.40.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1